Move case mappings to its own folder
