+++ x-pack/plugins/triggers_actions_ui/public/application/components/builtin_action_types/case_mappings/translations.ts
