+++ x-pack/plugins/triggers_actions_ui/public/application/components/builtin_action_types/jira/jira_connectors.tsx
@@ -16,10 +16,13 @@ import {
 
 import { isEmpty } from 'lodash';
 import { ActionConnectorFieldsProps } from '../../../../types';
+import { CasesConfigurationMapping } from '../case_mappings/types';
+import { FieldMapping } from '../case_mappings/field_mapping';
+import { createDefaultMapping } from '../case_mappings/utils';
+
 import * as i18n from './translations';
-import { JiraActionConnector, CasesConfigurationMapping } from './types';
+import { JiraActionConnector } from './types';
 import { connectorConfiguration } from './config';
-import { FieldMapping } from './case_mappings/field_mapping';
 
 const JiraConnectorFields: React.FC<ActionConnectorFieldsProps<JiraActionConnector>> = ({
   action,
@@ -192,7 +195,7 @@ const JiraConnectorFields: React.FC<ActionConnectorFieldsProps<JiraActionConnect
             <EuiFlexItem data-test-subj="case-jira-mappings">
               <FieldMapping
                 disabled={true}
-                connectorActionTypeId={connectorConfiguration.id}
+                connectorConfiguration={connectorConfiguration}
                 mapping={mapping as CasesConfigurationMapping[]}
                 onChangeMapping={handleOnChangeMappingConfig}
               />
@@ -203,16 +206,6 @@ const JiraConnectorFields: React.FC<ActionConnectorFieldsProps<JiraActionConnect
     </>
   );
 };
-
-export const createDefaultMapping = (fields: Record<string, any>): CasesConfigurationMapping[] =>
-  Object.keys(fields).map(
-    (key) =>
-      ({
-        source: fields[key].defaultSourceField,
-        target: key,
-        actionType: fields[key].defaultActionType,
-      } as CasesConfigurationMapping)
-  );
 
 // eslint-disable-next-line import/no-default-export
 export { JiraConnectorFields as default };
