--- x-pack/plugins/triggers_actions_ui/public/application/components/builtin_action_types/servicenow/case_mappings/types.ts
@@ -1,16 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { ActionType } from '../../../../../types';
-
-export { ActionType };
-
-export interface ThirdPartyField {
-  label: string;
-  validSourceFields: string[];
-  defaultSourceField: string;
-  defaultActionType: string;
-}
