--- x-pack/plugins/triggers_actions_ui/public/application/components/builtin_action_types/servicenow/case_mappings/utils.ts
@@ -1,38 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-import { CasesConfigurationMapping } from '../types';
-
-export const setActionTypeToMapping = (
-  caseField: string,
-  newActionType: string,
-  mapping: CasesConfigurationMapping[]
-): CasesConfigurationMapping[] => {
-  const findItemIndex = mapping.findIndex((item) => item.source === caseField);
-
-  if (findItemIndex >= 0) {
-    return [
-      ...mapping.slice(0, findItemIndex),
-      { ...mapping[findItemIndex], actionType: newActionType },
-      ...mapping.slice(findItemIndex + 1),
-    ];
-  }
-
-  return [...mapping];
-};
-
-export const setThirdPartyToMapping = (
-  caseField: string,
-  newThirdPartyField: string,
-  mapping: CasesConfigurationMapping[]
-): CasesConfigurationMapping[] =>
-  mapping.map((item) => {
-    if (item.source !== caseField && item.target === newThirdPartyField) {
-      return { ...item, target: 'not_mapped' };
-    } else if (item.source === caseField) {
-      return { ...item, target: newThirdPartyField };
-    }
-    return item;
-  });
