+++ x-pack/plugins/triggers_actions_ui/public/application/components/builtin_action_types/servicenow/types.ts
@@ -4,6 +4,8 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { CasesConfigurationMapping } from '../case_mappings/types';
+
 export interface ServiceNowActionConnector {
   config: ServiceNowConfig;
   secrets: ServiceNowSecrets;
@@ -36,11 +38,4 @@ interface ServiceNowConfig {
 interface ServiceNowSecrets {
   username: string;
   password: string;
-}
-
-// to remove
-export interface CasesConfigurationMapping {
-  source: string;
-  target: string;
-  actionType: string;
 }
